--- HunkyDoryCode.java
@@ -1,16 +1,0 @@
-public class MainEmpleado {
-    public static void main(String[] args){
-        // Todo, adiciona tú código de aplicación
-        String Nombres, Apellidos, Id, cargo;
-        int edad;
-        System.out.println("Ingresar los nombres: ");
-
-
-        //Instanciar la clase1 Empresa
-        Empresa entidad1 = new Empresa("valor que se captura con el scanner, como variable", "Direccion de la empresa", 6751389, 848950294);
-
-        //Instanciar la clase2 movimientodinero:
-        movimientoDinero entradaysalida = new movimientoDinero(0, 0, 0, "", "");
-
-    }
-}
