--- MainEmpleado.java
@@ -1,9 +1,0 @@
-public class MainEmpleado {
-    public static void main(String[] args){
-        // Todo, adiciona tú código de aplicación
-        String Nombres, Apellidos, Id, cargo;
-        int edad;
-        System.out.println("Ingresar los nombres: ");
-
-    }
-}
