--- archivo1.txt
@@ -1,1 +1,0 @@
-Este es mi primer commit, Esteban Castano para HunkyDoryCode
